+++ myRayTracing/src/shapes.jl
@@ -233,7 +233,7 @@ function ray_intersection(shape::Plane, r::Ray)
 
     inv_r = inverse(shape.T)(r)
 
-    if inv_r.dir.z == 0
+    if abs(inv_r.dir.z) < 1e-6
         return nothing
     end
  
@@ -259,16 +259,19 @@ end
 function quick_ray_intersection(shape, r)
     given a plane and a ray, it returns true/false if there is/isn't intersection
 """
-function quick_ray_intersection(shape::Plane, r::Ray)::Bool
+function quick_ray_intersection(shape::Plane, r::Ray)
     inv_r = inverse(shape.T)(r)
 
-    if abs(inv_r.dir.z) < 1e-5
+    if abs(inv_r.dir.z) < 1e-6
         return false
+    else 
+        t_hit = - inv_r.origin.z / inv_r.dir.z
+        if t_hit < inv_r.tmin || t_hit > inv_r.tmax
+            return false
+        else
+            return true
+        end
     end
-
-    t = -inv_r.origin.z / inv_r.dir.z
-
-    return inv_r.tmin < t < inv_r.tmax
 end
 
 #################################################################################
@@ -418,6 +421,9 @@ struct Box <: Shape
     material::Material
 
     function Box(X::Float64, Y::Float64, Z::Float64, T::Transformation=Transformation(Matrix{Float64}(I(4))), material::Material=Material())
+        if X <= 0.0 || Y <= 0.0 || Z <= 0.0
+            throw(ArgumentError("Box.X, .Y and .Z must be positives"))
+        end
         new(X, Y, Z, T, material)
     end
 end
@@ -467,7 +473,7 @@ function all_ray_intersection(box::Box, r::Ray)
 
     inv_r = inverse(box.T)(r)
 
-    if inv_r.dir.x == 0.0
+    if abs(inv_r.dir.x) < 1e-6
         if (inv_r.origin.x < 0.0 || inv_r.origin.x > box.X)
             return hits
         else
@@ -479,7 +485,7 @@ function all_ray_intersection(box::Box, r::Ray)
     end
 
 
-    if inv_r.dir.y == 0.0
+    if abs(inv_r.dir.y) < 1e-6
         if (inv_r.origin.y < 0.0 || inv_r.origin.y > box.Y)
             return hits
         else
@@ -491,7 +497,7 @@ function all_ray_intersection(box::Box, r::Ray)
     end
 
 
-    if inv_r.dir.z == 0.0
+    if abs(inv_r.dir.z) < 1e-6
         if (inv_r.origin.z < 0.0 || inv_r.origin.z > box.Z)
             return hits
         else
@@ -505,7 +511,7 @@ function all_ray_intersection(box::Box, r::Ray)
     t_1 = max(int_x[1], int_y[1], int_z[1])
     t_2 = min(int_x[2], int_y[2], int_z[2])
 
-    if (t_2-t_1) < 1e-6 #check both t2>=t1 and  t2!=t1
+    if (t_2-t_1) < 1e-6 #check both t2>=t1 and t2!=t1
         return hits
     end
 
@@ -567,7 +573,7 @@ function quick_ray_intersection(box::Box, r::Ray)
 
     inv_r = inverse(box.T)(r)
 
-    if inv_r.dir.x == 0.0
+    if abs(inv_r.dir.x) < 1e-6
         if (inv_r.origin.x < 0.0 || inv_r.origin.x > box.X)
             return false
         else
@@ -579,7 +585,7 @@ function quick_ray_intersection(box::Box, r::Ray)
     end
 
 
-    if inv_r.dir.y == 0.0
+    if abs(inv_r.dir.y) < 1e-6
         if (inv_r.origin.y < 0.0 || inv_r.origin.y > box.Y)
             return false
         else
@@ -591,7 +597,7 @@ function quick_ray_intersection(box::Box, r::Ray)
     end
 
 
-    if inv_r.dir.z == 0.0
+    if abs(inv_r.dir.z) < 1e-6
         if (inv_r.origin.z < 0.0 || inv_r.origin.z > box.Z)
             return false
         else
@@ -605,14 +611,517 @@ function quick_ray_intersection(box::Box, r::Ray)
     t_1 = max(int_x[1], int_y[1], int_z[1])
     t_2 = min(int_x[2], int_y[2], int_z[2])
 
-    if ((t_1 > inv_r.tmin) && (t_1 < inv_r.tmax)) || ((t_2 > inv_r.tmin) && (t_2 < inv_r.tmax))
-        true  
+    if ((t_2-t_1) > 1e-6) && ( ((t_1 > inv_r.tmin) && (t_1 < inv_r.tmax)) || ((t_2 > inv_r.tmin) && (t_2 < inv_r.tmax)) )
+        return true  
     else
         return false
     end
 end
 
 
+####################################################################################################
+
+#CYLINDER STRUCT
+"""
+struct Cylinder <: Shape
+    Creates a cylinder. It has a base on the xy plane centered in the origin and his axis along z axis.
+
+    R::Float64 --> Radius of the cylinder's base
+    H::Float64 --> Height of the cylinder
+    T::Transformation --> The transformation applied to the cylinder
+    material::Material --> the material of the cylinder
+
+    function Cylinder(R::Float64, H::Float64, T::Transformation=Transformation(Matrix{Float64}(I(4))), material::Material=Material())
+        new(R, H, T, material)
+    end
+end
+"""
+struct Cylinder <: Shape
+ 
+    R::Float64
+    H::Float64
+    T::Transformation
+    material::Material
+
+    function Cylinder(R::Float64, H::Float64, T::Transformation=Transformation(Matrix{Float64}(I(4))), material::Material=Material())
+        if R <= 0.0 || H <= 0.0
+            throw(ArgumentError("Cylinder's Radius and Height must be positives"))
+        end
+        new(R, H, T, material)
+    end
+end
+
+"""
+function _shape_normal(cyl, p, r)
+    it returns the normal to the cylinder surface in a given point, chosen in order to have the opposite direction of the incoming ray.
+"""
+function _shape_normal(cyl::Cylinder, r::Ray, p::Point)
+    epsilon = 1e-6
+
+    if abs(p.z - 0.0) < epsilon || abs(p.z - cyl.H) < epsilon
+        n = Normal(0.0, 0.0, 1.0)
+    else
+        n = normalize(Normal(p.x, p.y, 0.0))
+    end
+
+    if (n * r.dir) < 0
+        return n
+    else
+        return neg(n)
+    end
+end
+
+"""
+function _xyz_to_uv(cyl, p)
+    given a point on the cylinder, it returns its (u,v) 2D form.
+"""
+function _xyz_to_uv(cyl::Cylinder, p::Point)
+    epsilon = 1e-6 
+
+    if abs(p.z - 0.0) < epsilon || abs(p.z - cyl.H) < epsilon
+        u = (p.x + cyl.R) / (2 * cyl.R)
+        v = (p.y + cyl.R) / (2 * cyl.R)
+    else
+        u = ( atan(p.y, p.x) + π ) / (2π)
+        v = p.z / cyl.H 
+    end
+    return Vec2d(u, v)
+end
+
+"""
+function all_ray_intersection(cyl, r)
+    given a cylinder and a ray, it returns all the HitRecords for the intersection between the ray and the cylinder
+"""
+function all_ray_intersection(cyl::Cylinder, r::Ray)
+    hits = Vector{Tuple{HitRecord, HitRecord}}()
+
+    inv_r = inverse(cyl.T)(r)
+
+    if abs(inv_r.dir.x) < 1e-6 && abs(inv_r.dir.y) < 1e-6
+        if √(inv_r.origin.x^2 + inv_r.origin.y^2) >= cyl.R
+            return hits
+        else
+            int_xy = (-Inf,Inf)
+        end
+    else
+        a = inv_r.dir.x^2 + inv_r.dir.y^2
+        b = inv_r.origin.x * inv_r.dir.x + inv_r.origin.y * inv_r.dir.y #tecnically is b/2, but we will use delta/4
+        c = inv_r.origin.x^2 + inv_r.origin.y^2 - cyl.R^2
+        delta = b*b - a*c #it's delta/4
+        if delta <= 0
+            return hits
+        else
+            sqrt_delta = √(delta)
+            int_xy = ( (-b-sqrt_delta)/a , (-b+sqrt_delta)/a)
+        end
+    end
+
+    if abs(inv_r.dir.z) < 1e-6
+        if (inv_r.origin.z < 0.0 || inv_r.origin.z > cyl.H)
+            return hits
+        else
+            int_z = (-Inf,Inf)
+        end
+    else
+        int_z = (min(-inv_r.origin.z/inv_r.dir.z,(cyl.H-inv_r.origin.z)/inv_r.dir.z), 
+                 max(-inv_r.origin.z/inv_r.dir.z,(cyl.H-inv_r.origin.z)/inv_r.dir.z))
+    end
+
+    t_1 = max(int_xy[1], int_z[1])
+    t_2 = min(int_xy[2], int_z[2])
+
+    if (t_2-t_1) < 1e-6 #check both t2>=t1 and t2!=t1
+        return hits
+    end
+
+    if t_1 < inv_r.tmax
+        point_hit_1 = at(inv_r, t_1)
+        hit_1 = HitRecord(cyl.T(point_hit_1), #hitted point in the world
+                          cyl.T(_shape_normal(cyl, inv_r, point_hit_1)), #normal at the surface in the world
+                          (_xyz_to_uv(cyl, point_hit_1)), #(u,v) vec hitted on the surface
+                          t_1, #t
+                          r, #ray
+                          cyl #cylinder
+                          )
+    
+    end
+
+    if t_2 < inv_r.tmax
+        point_hit_2 = at(inv_r, t_2)
+        hit_2 = HitRecord(cyl.T(point_hit_2), #hitted point in the world
+                          cyl.T(_shape_normal(cyl, inv_r, point_hit_2)), #normal at the surface in the world
+                          (_xyz_to_uv(cyl, point_hit_2)), #(u,v) vec hitted on the surface
+                          t_2, #t
+                          r, #ray
+                          cyl #cylinder
+                          )
+             
+    end
+    
+    push!(hits, (hit_1, hit_2))
+
+    return hits
+
+end
+
+"""
+function ray_intersection(cyl, r)
+    given a cylinder and a ray, it returns the HitRecord for the first intersection between the cylinder and the box
+"""
+function ray_intersection(cyl::Cylinder, r::Ray)
+    all_hits = all_ray_intersection(cyl, r)
+
+    if !isempty(all_hits)
+        first_hit_index = findfirst(hit -> hit.t > r.tmin, all_hits[1])
+
+        if first_hit_index !== nothing
+            return all_hits[1][first_hit_index]
+        else
+            return nothing
+        end 
+    else
+        return nothing
+    end
+end
+
+"""
+function quick_ray_intersection(cyl, r)
+    given a cylinder and a ray, it returns whether the intersection between the ray and the ray and the cylinder happens or not
+"""
+function quick_ray_intersection(cyl::Cylinder, r::Ray)
+    inv_r = inverse(cyl.T)(r)
+
+    if abs(inv_r.dir.x) < 1e-6 && abs(inv_r.dir.y) < 1e-6
+        if √(inv_r.dir.x^2 + inv_r.dir.y^2) >= cyl.R
+            return false
+        else
+            int_xy = (-Inf,Inf)
+        end
+    else
+        a = inv_r.dir.x^2 + inv_r.dir.y^2
+        b = inv_r.origin.x * inv_r.dir.x + inv_r.origin.y * inv_r.dir.y #tecnically is b/2, but we will use delta/4
+        c = inv_r.origin.x^2 + inv_r.origin.y^2 - cyl.R^2
+        delta = b*b - a*c #it's delta/4
+        if delta <= 0
+            return false
+        else
+            sqrt_delta = √(delta)
+            int_xy = ( (-b-sqrt_delta)/a , (-b+sqrt_delta)/a)
+        end
+    end
+
+    if abs(inv_r.dir.z) < 1e-6
+        if (inv_r.origin.z < 0.0 || inv_r.origin.z > cyl.H)
+            return false
+        else
+            int_z = (-Inf,Inf)
+        end
+    else
+        int_z = (min(-inv_r.origin.z/inv_r.dir.z,(cyl.H-inv_r.origin.z)/inv_r.dir.z), 
+                 max(-inv_r.origin.z/inv_r.dir.z,(cyl.H-inv_r.origin.z)/inv_r.dir.z))
+    end
+
+    t_1 = max(int_xy[1], int_z[1])
+    t_2 = min(int_xy[2], int_z[2])
+
+    if ((t_2-t_1) > 1e-6) && ( ((t_1 > inv_r.tmin) && (t_1 < inv_r.tmax)) || ((t_2 > inv_r.tmin) && (t_2 < inv_r.tmax)) )
+        return true  
+    else
+        return false
+    end
+
+end
+
+###################################################################################################
+#CONE STRUCT
+
+"""
+struct Cone <: Shape
+    Creates a cone. It has a base on the xy plane centered in the origin and its axis along the z axis.
+
+    R::Float64 --> Radius of the cone's base
+    H::Float64 --> Height of the cone
+    T::Transformation --> The transformation applied to the cone
+    material::Material --> the material of the cone
+
+    function Cone(R::Float64, H::Float64, T::Transformation=Transformation(Matrix{Float64}(I(4))), material::Material=Material())
+        new(R, H, T, material)
+    end
+end
+"""
+struct Cone <: Shape
+    R::Float64
+    H::Float64
+    T::Transformation
+    material::Material
+
+    function Cone(R::Float64, H::Float64, T::Transformation=Transformation(Matrix{Float64}(I(4))), material::Material=Material())
+        if R <= 0.0 || H <= 0.0
+            throw(ArgumentError("Cone's Radius and Height must be positives"))
+        end
+        new(R, H, T, material)
+    end
+end
+
+"""
+function _shape_normal(cone, p, r)
+    it returns the normal to the cone surface in a given point, chosen in order to have the opposite direction of the incoming ray.
+"""
+function _shape_normal(cone::Cone, r::Ray, p::Point)
+    epsilon = 1e-6
+
+    if abs(p.z) < epsilon || abs(p.z - cone.H) < epsilon
+        n = Normal(0.0, 0.0, 1.0)
+    else
+        n = normalize(Normal(p.x, p.y, cone.R^2/cone.H * (1 - p.z/cone.H)))
+    end
+
+    if (n * r.dir) < 0
+        return n
+    else
+        return neg(n)
+    end
+end
+
+"""
+function _xyz_to_uv(cone, p)
+    given a point on the cone, it returns its (u,v) 2D form.
+"""
+function _xyz_to_uv(cone::Cone, p::Point)
+    epsilon = 1e-6 
+
+    if abs(p.z) < epsilon
+        u = (p.x + cone.R) / (2 * cone.R)
+        v = (p.y + cone.R) / (2 * cone.R)
+    else
+        u = ( atan(p.y, p.x) + π ) / (2π)
+        v = p.z / cone.H 
+    end
+    return Vec2d(u, v)
+end
+
+"""
+function all_ray_intersection(cone, r)
+    given a cone and a ray, it returns all the HitRecords for the intersection between the ray and the cone
+"""
+function all_ray_intersection(cone::Cone, r::Ray)
+    hits = Vector{Tuple{HitRecord, HitRecord}}()
+
+    inv_r = inverse(cone.T)(r)
+
+    if abs(inv_r.dir.x) < 1e-6 && abs(inv_r.dir.y) < 1e-6
+        r_xy = √(inv_r.origin.x^2 + inv_r.origin.y^2)
+
+        if r_xy >= cone.R
+            return hits
+        elseif inv_r.dir.z > 0.0
+            z_hit = cone.H * (1 - r_xy/cone.R)
+            int_xy = (-Inf, (z_hit-inv_r.origin.z)/inv_r.dir.z)
+        else
+            z_hit = cone.H * (1 - r_xy/cone.R)
+            int_xy = ((z_hit-inv_r.origin.z)/inv_r.dir.z, Inf)
+        end
+
+    else
+
+        a = inv_r.dir.x^2 + inv_r.dir.y^2 - (cone.R^2 / cone.H^2) * inv_r.dir.z^2
+        b = inv_r.origin.x * inv_r.dir.x + inv_r.origin.y * inv_r.dir.y + (cone.R^2 / cone.H^2) * inv_r.dir.z * (cone.H - inv_r.origin.z) #tecnically is b/2, but we will use delta/4
+        c = inv_r.origin.x^2 + inv_r.origin.y^2 - (cone.R^2 / cone.H^2) * (cone.H - inv_r.origin.z)^2  
+
+        if abs(a) < 1e-6
+            t_single_hit = - c / 2b
+            test_point = at(inv_r, t_single_hit)
+            if test_point.z > cone.H
+                return hits
+            elseif inv_r.dir.z < 0.0
+                int_xy = (t_single_hit, Inf)
+            else
+                int_xy = (-Inf, t_single_hit)
+            end        
+        else
+
+            delta = b*b - a*c #it's delta/4
+
+            if delta <= 0
+                return hits
+            else
+                sqrt_delta = √(delta)
+                t_xy_1 = min( (-b-sqrt_delta)/a, (-b+sqrt_delta)/a )
+                t_xy_2 = max( (-b-sqrt_delta)/a, (-b+sqrt_delta)/a )
+
+                test_point_1 = at(inv_r, t_xy_1)
+                test_point_2 = at(inv_r, t_xy_2)
+
+                if test_point_1.z < cone.H
+                    if test_point_2.z < cone.H
+                        int_xy = (t_xy_1, t_xy_2)
+                    else 
+                        int_xy = (-Inf, t_xy_1)
+                    end
+                elseif test_point_2.z < cone.H
+                    int_xy = (t_xy_2, Inf)
+                else
+                    return hits
+                end
+            end 
+        end    
+    end
+
+    if abs(inv_r.dir.z) < 1e-6
+        if (inv_r.origin.z < 0.0 || inv_r.origin.z > cone.H)
+            return hits
+        else
+            int_z = (-Inf,Inf)
+        end
+    else
+        int_z = (min(-inv_r.origin.z/inv_r.dir.z,(cone.H-inv_r.origin.z)/inv_r.dir.z), 
+                 max(-inv_r.origin.z/inv_r.dir.z,(cone.H-inv_r.origin.z)/inv_r.dir.z))
+    end
+
+    t_1 = max(int_xy[1], int_z[1])
+    t_2 = min(int_xy[2], int_z[2])
+
+
+    if (t_2-t_1) < 1e-6 #check both t2>=t1 and t2!=t1
+        return hits
+    end
+
+    if t_1 < inv_r.tmax
+        point_hit_1 = at(inv_r, t_1)
+        hit_1 = HitRecord(cone.T(point_hit_1), #hitted point in the world
+                          cone.T(_shape_normal(cone, inv_r, point_hit_1)), #normal at the surface in the world
+                          (_xyz_to_uv(cone, point_hit_1)), #(u,v) vec hitted on the surface
+                          t_1, #t
+                          r, #ray
+                          cone #cone
+                          )
+    
+    end
+
+    if t_2 < inv_r.tmax
+        point_hit_2 = at(inv_r, t_2)
+        hit_2 = HitRecord(cone.T(point_hit_2), #hitted point in the world
+                          cone.T(_shape_normal(cone, inv_r, point_hit_2)), #normal at the surface in the world
+                          (_xyz_to_uv(cone, point_hit_2)), #(u,v) vec hitted on the surface
+                          t_2, #t
+                          r, #ray
+                          cone #cone
+                          )
+             
+    end
+    
+    push!(hits, (hit_1, hit_2))
+
+    return hits
+
+end
+
+"""
+function ray_intersection(cone, r)
+    given a cone and a ray, it returns the HitRecord for the first intersection between the cone and the box
+"""
+function ray_intersection(cone::Cone, r::Ray)
+    all_hits = all_ray_intersection(cone, r)
+
+    if !isempty(all_hits)
+        first_hit_index = findfirst(hit -> hit.t > r.tmin, all_hits[1])
+
+        if first_hit_index !== nothing
+            return all_hits[1][first_hit_index]
+        else
+            return nothing
+        end 
+    else
+        return nothing
+    end
+end
+
+"""
+function quick_ray_intersection(cone, r)
+    given a cone and a ray, it returns whether the intersection between the ray and the ray and the cone happens or not
+"""
+function quick_ray_intersection(cone::Cone, r::Ray)
+    inv_r = inverse(cone.T)(r)
+
+    if abs(inv_r.dir.x) < 1e-6 && abs(inv_r.dir.y) < 1e-6
+        r = √(inv_r.origin.x^2 + inv_r.origin.y^2)
+
+        if r >= cone.R
+            return false
+        elseif inv_r.dir.z > 0.0
+            z_hit = cone.H * (1 - r/cone.R)
+            int_xy = (-Inf, (z_hit-inv_r.origin.z)/inv_r.dir.z)
+        else
+            z_hit = cone.H * (1 - r/cone.R)
+            int_xy = ((z_hit-inv_r.origin.z)/inv_r.dir.z, Inf)
+        end
+
+    else
+
+        a = inv_r.dir.x^2 + inv_r.dir.y^2 - (cone.R^2 / cone.H^2) * inv_r.dir.z^2
+        b = inv_r.origin.x * inv_r.dir.x + inv_r.origin.y * inv_r.dir.y + (cone.R^2 / cone.H^2) * inv_r.dir.z * (cone.H - inv_r.origin.z) #tecnically is b/2, but we will use delta/4
+        c = inv_r.origin.x^2 + inv_r.origin.y^2 - (cone.R^2 / cone.H^2) * (cone.H - inv_r.origin.z)^2  
+
+        if abs(a) < 1e-6
+            t_single_hit = - c / b
+            test_point = at(inv_r, t_single_hit)
+            if test_point.z > cone.H
+                return false
+            elseif inv_r.dir.z < 0.0
+                int_xy = (t_single_hit, Inf)
+            else
+                int_xy = (-Inf, t_single_hit)
+            end
+        end
+
+        delta = b*b - a*c #it's delta/4
+
+        if delta <= 0
+            return false
+        else
+            sqrt_delta = √(delta)
+            t_xy_1 = min( (-b-sqrt_delta)/a, (-b+sqrt_delta)/a )
+            t_xy_2 = max( (-b-sqrt_delta)/a, (-b+sqrt_delta)/a )
+
+            test_point_1 = at(inv_r, t_xy_1)
+            test_point_2 = at(inv_r, t_xy_2)
+
+            if test_point_1.z < cone.H
+                if test_point_2.z < cone.H
+                    int_xy = (t_xy_1, t_xy_2)
+                else 
+                    int_xy = (-Inf, t_xy_1)
+                end
+            elseif test_point_2.z < cone.H
+                int_xy = (t_xy_2, Inf)
+            else
+                return false
+            end
+        end     
+    end
+
+    if abs(inv_r.dir.z) < 1e-6
+        if (inv_r.origin.z < 0.0 || inv_r.origin.z > cone.H)
+            return false
+        else
+            int_z = (-Inf,Inf)
+        end
+    else
+        int_z = (min(-inv_r.origin.z/inv_r.dir.z,(cone.H-inv_r.origin.z)/inv_r.dir.z), 
+                 max(-inv_r.origin.z/inv_r.dir.z,(cone.H-inv_r.origin.z)/inv_r.dir.z))
+    end
+
+    t_1 = max(int_xy[1], int_z[1])
+    t_2 = min(int_xy[2], int_z[2])
+
+    if ((t_2-t_1) > 1e-6) && ( ((t_1 > inv_r.tmin) && (t_1 < inv_r.tmax)) || ((t_2 > inv_r.tmin) && (t_2 < inv_r.tmax)) )
+        return true  
+    else
+        return false
+    end
+end
+
+####################################################################################################
 #HIT RECORD STRUCT
 
 """
@@ -851,6 +1360,25 @@ function ray_intersection(u_shape::union_shape, r::Ray)
     end
 end
 
+"""
+function quick_ray_intersection(u_shape::union_shape, r::Ray)
+    Given a union of shapes and a ray, it returns whether the intersection between the ray and the shape happens or not
+"""
+function quick_ray_intersection(u_shape::union_shape, r::Ray)
+    all_hits_intervals = all_ray_intersection(u_shape, r)
+
+    if !isempty(all_hits_intervals)
+        for (entry_hit, exit_hit) in all_hits_intervals
+            if entry_hit.t > r.tmin || exit_hit.t > r.tmin
+                return true
+            end
+        end
+        return false
+    else
+        return false
+    end
+end
+
 #INTERSECTION
 """
 new shape type for intersection in CSG:
@@ -909,6 +1437,25 @@ function ray_intersection(i_shape::intersec_shape, r::Ray)
     end
 end
 
+"""
+function quick_ray_intersection(i_shape::intersec_shape, r::Ray)
+    Given an intersection of shapes and a ray, it returns whether the intersection between the ray and the shape happens or not
+"""
+function quick_ray_intersection(i_shape::intersec_shape, r::Ray)
+    all_hits_intervals = all_ray_intersection(i_shape, r)
+
+    if !isempty(all_hits_intervals)
+        for (entry_hit, exit_hit) in all_hits_intervals
+            if entry_hit.t > r.tmin || exit_hit.t > r.tmin
+                return true
+            end
+        end
+        return false
+    else
+        return false
+    end
+end
+
 #DIFFERENCE
 """
 new shape type for difference in CSG:
@@ -964,5 +1511,24 @@ function ray_intersection(d_shape::diff_shape, r::Ray)
         return nothing
     else
         return nothing
+    end
+end
+
+"""
+function quick_ray_intersection(d_shape::diff_shape, r::Ray)
+    Given a difference of shapes and a ray, it returns whether the intersection between the ray and the shape happens or not
+"""
+function quick_ray_intersection(d_shape::diff_shape, r::Ray)
+    all_hits_intervals = all_ray_intersection(d_shape, r)
+
+    if !isempty(all_hits_intervals)
+        for (entry_hit, exit_hit) in all_hits_intervals
+            if entry_hit.t > r.tmin || exit_hit.t > r.tmin
+                return true
+            end
+        end
+        return false
+    else
+        return false
     end
 end